+++ actionpack/lib/action_dispatch/middleware/exception_wrapper.rb
@@ -17,8 +17,8 @@ module ActionDispatch
       'ActionDispatch::ParamsParser::ParseError'      => :bad_request,
       'ActionController::BadRequest'                  => :bad_request,
       'ActionController::ParameterMissing'            => :bad_request,
-      'Rack::Utils::ParameterTypeError'               => :bad_request,
-      'Rack::Utils::InvalidParameterError'            => :bad_request
+      'Rack::QueryParser::ParameterTypeError'         => :bad_request,
+      'Rack::QueryParser::InvalidParameterError'      => :bad_request
     )
 
     cattr_accessor :rescue_templates